Create a new file so that a user does not need pydantic
